move objects from Objects/.../object to src/Object
