+++ src/Objects/Saveable.java
@@ -1,13 +1,19 @@
-package Objects;
+package src.main.java.objects;
 
 import java.util.*;
 
+//Class to generate a unique ID
 public class Saveable {
-    public String ID;
-    public String get_id(){return this.ID;}
-    public Hashtable<String, String> to_json(){
-        Hashtable<String, String> result = new Hashtable<String, String>();
-        result.put("id", this.ID);
-        return result;
+    private final String ID;
+    
+    //Generating a new unique ID
+    public Saveable(){
+        this.ID = UUID.randomUUID().toString();
     }
+    
+    public Saveable(String id){
+        this.ID = id;
+    }
+    
+    public String get_id(){return this.ID;}
 }
