+++ src/Objects/Score.java
@@ -1,22 +1,22 @@
-package Objects;
+package src.main.java.objects;
 
-import java.util.*;
-
-public class Score extends Saveable{
+//Score class that can used to store and update a player's score
+public class Score extends Saveable {
     int cur_score;
     
     public Score(){
-        this.ID = UUID.randomUUID().toString();
+        super();
         this.cur_score = 0;
     }
     
     public Score(int score){
-        this.ID = UUID.randomUUID().toString();
+    	super();
         this.cur_score = score;
     }
     
+    //This is done so that the score can be initialized based on the player's ID
     public Score(String id, int score){
-        this.ID = id;
+    	super(id);
         this.cur_score = score;
     }
     
@@ -24,10 +24,8 @@ public class Score extends Saveable{
         return this.cur_score;
     }
     
+    //To update the score depending on the player's move
     public int update(int change){
-        //The score will be updated based on the rules that will be provided by Team 4
-        //Iterate through the rule list and change the score based on each rule
-        // int new_score = 1;//Calculate based on rules and update
         this.cur_score = change;
         return this.cur_score;
     }
