+++ src/Objects/Token.java
@@ -1,8 +1,8 @@
-package Objects;
+package src.main.java.objects;
 
 import java.util.*;
 
-public class Token extends Savable implements IToken {
+public class Token extends Saveable implements IToken {
 	
 	private Board board;
 	private ArrayList<Player> players;
